♻️ Refactor: UI 컴포넌트 대소문자 통일
refactor: UI 컴포넌트 대소문자 통일
