+++ src/features/schedule/widgets/DateBottomSheet.tsx
@@ -2,6 +2,7 @@ import { DatePickerDrawer, DateSelector } from "@/shared/ui";
 
 const handleDateConfirm = (date: Date) => {
   // setBirthday(formatDateKorean(date));
+  console.log("date", date);
 };
 
 export const DateBottomSheet = () => {
