+++ src/shared/ui/index.ts
@@ -11,9 +11,9 @@ export * from "./navbar";
 export * from "./progressheader";
 export * from "./switch";
 export * from "./textinput";
-export * from "./Topbar";
-export * from "./TimeSelector";
-export * from "./DateSelector";
+export * from "./topbar";
+export * from "./timeSelectBox";
+export * from "./dateSelectBox";
 export * from "./shadcnInput";
 export * from "./textarea";
 export * from "./timepicker";
